Upgrade pegasus example from two to arbitrary number of drones  equally spaced.
